backend/requirements.txt
asgiref==3.7.2; python_version >= "3.8"
authlib==1.2.0
certifi==2023.5.7; python_version >= "3.7"
cffi==1.15.1; python_version >= "3.7"
charset-normalizer==3.1.0; python_full_version >= "3.7.0" and python_version >= "3.7"
cryptography==41.0.1; python_version >= "3.7"
django-cors-headers==4.0.0; python_version >= "3.7"
django==4.2.2; python_version >= "3.8"
djangorestframework-simplejwt==5.2.2; python_version >= "3.7"
djangorestframework==3.14.0; python_version >= "3.6"
ecdsa==0.18.0; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
gunicorn==20.1.0; python_version >= "3.5"
idna==3.4; python_version >= "3.7"
pyasn1==0.5.0; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4" or python_version >= "3.6" and python_version < "4" and python_full_version >= "3.6.0"
pycparser==2.21; python_version >= "3.7" and python_full_version < "3.0.0" or python_version >= "3.7" and python_full_version >= "3.4.0"
pyjwt==2.7.0; python_version >= "3.7"
python-dotenv==1.0.0; python_version >= "3.8"
python-jose==3.3.0
pytz==2023.3; python_version >= "3.7"
requests==2.31.0; python_version >= "3.7"
rsa==4.9; python_version >= "3.6" and python_version < "4"
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
sqlparse==0.4.4; python_version >= "3.8"
typing-extensions==4.6.3; python_version < "3.11" and python_version >= "3.8"
tzdata==2023.3; sys_platform == "win32" and python_version >= "3.8"
urllib3==2.0.2; python_version >= "3.7"
